Update PDC-831_DQ-EncounterFrequency.json
removed reference to active flag as this is confusing to me - all patients in the endpoint must have been flagged active to be present in the E2E extract from the vendor side.  The denominator is counting all patients.  The code didn't reflect the description - removed reference to prescription as it wasn't being looked for.......this description more accurately reflects the query code.